Se separan en 2 custom hooks, dependiendo si se llaman datos o un documento entero
