+++ src/pages/ExaDetail/ExaDetail.js
@@ -1,10 +1,9 @@
 import { useContext, useEffect, useState } from 'react'
 import { Navigate, useParams, Link } from 'react-router-dom'
-
 // Librerías
 import CircularProgress from '@mui/material/CircularProgress'
 // Custom Hooks
-import { useGetCases } from '../../customHooks/useGetData'
+import { useGetCases } from '../../customHooks/documentHook'
 // Context
 import { AuthContext } from '../../context/authContext'
 // Firebase
@@ -36,37 +35,29 @@ const ExaDetail = () => {
         {exa.toUpperCase()} - {userCases[0]?.nombre}
       </h2>
       {loading
-        ? (<CircularProgress />)
-        : (
-          <section className='exa-detail'>
-            <ul>
-              {userCases.map(c => {
-                const {
-                  numeroCaso,
-                  date,
-                  motivoConsulta,
-                  puntoATrabajar,
-                  comentarioGestion,
-                  id
-                } = c
-                return (
-                  <li className='' key={c.id}>
-                    <p>{numeroCaso}</p>
-                    <p>{date}</p>
-                    <p>{motivoConsulta}</p>
-                    {puntoATrabajar !== '-' && <p>{puntoATrabajar}</p>}
-                    <p className='commentary'>
-                      {comentarioGestion.substring(0, 50)}...
-                    </p>
-                    <Link to={`/monitoreo/${id}`}>
-                      <FeedIcon color='primary' fontSize='large' />
-                    </Link>
-                  </li>
-                )
-              })}
-            </ul>
-          </section>
-          )}
+        ? <CircularProgress />
+        : <section className='exa-detail'>
+          <ul>
+            {userCases.map(c => {
+              const { numeroCaso, date, motivoConsulta, puntoATrabajar, comentarioGestion, id } = c
+              return (
+                <li className='' key={c.id}>
+                  <p>{numeroCaso}</p>
+                  <p>{date}</p>
+                  <p>{motivoConsulta}</p>
+                  {puntoATrabajar !== '-' && <p>{puntoATrabajar}</p>}
+                  <p className='commentary'>
+                    {comentarioGestion.substring(0, 50)}...
+                  </p>
+                  <Link to={`/monitoreo/${id}`}>
+                    <FeedIcon color='primary' fontSize='large' />
+                  </Link>
+                </li>
+              )
+            })}
+          </ul>
+        </section>
+      }
     </>
   )
 }
